+++ src/infrastructure/transaction_bundle.rs
@@ -1,4 +1,5 @@
-use crate::{transaction_store::TransactionData, txn_sender::TxnSender};
+use super::txn_sender::TxnSender;
+use crate::storage::transaction_store::TransactionData;
 use anyhow::anyhow;
 use cadence_macros::statsd_count;
 use dashmap::DashMap;