+++ src/infrastructure/txn_sender.rs
@@ -15,11 +15,8 @@ use tokio::{
 use tonic::async_trait;
 use tracing::{debug, error, warn};
 
-use crate::{
-    leader_tracker::LeaderTracker,
-    solana_rpc::SolanaRpc,
-    transaction_store::{get_signature, TransactionData, TransactionStore},
-};
+use super::{leader_tracker::LeaderTracker, solana_rpc::SolanaRpc};
+use crate::storage::transaction_store::{get_signature, TransactionData, TransactionStore};
 use solana_program_runtime::compute_budget::DEFAULT_INSTRUCTION_COMPUTE_UNIT_LIMIT;
 use solana_sdk::borsh0_10::try_from_slice_unchecked;
 use solana_sdk::compute_budget::ComputeBudgetInstruction;