+++ src/lib.rs
@@ -1,10 +1,8 @@
 pub mod errors;
-pub mod grpc_geyser;
-pub mod leader_tracker;
-pub mod metrics;
-pub mod rpc_server;
-pub mod solana_rpc;
-pub mod transaction_bundle;
-pub mod transaction_store;
-pub mod txn_sender;
 pub mod vendor;
+
+pub mod application;
+pub mod infrastructure;
+pub mod metrics;
+pub mod server;
+pub mod storage;